Finishing touches (comments + cleanup)
