--- Aural/PreferredDevices.swift
@@ -1,12 +1,0 @@
-import Foundation
-
-public class PreferredDevice {
-    
-    var name: String
-    var uid: String
-    
-    init(_ name: String, _ uid: String) {
-        self.name = name
-        self.uid = uid
-    }
-}
