+++ Aural/SoundPreferencesViewController.swift
@@ -261,3 +261,15 @@ class SoundPreferencesViewController: NSViewController, PreferencesViewProtocol 
         }
     }
 }
+
+// Encapsulates a user-preferred audio output device
+public class PreferredDevice {
+    
+    var name: String
+    var uid: String
+    
+    init(_ name: String, _ uid: String) {
+        self.name = name
+        self.uid = uid
+    }
+}
